backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
sqlparse
groq
python-dotenv
requests
autopep8